HEEDLS-492 Update AllDelegateItems so fetching is done in controller
